pyproject.toml
[project]
name = "viz-art"
version = "0.1.0"
description = "Lightweight library for building multi-stage vision processing pipelines"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pydantic>=2.0",
    "omegaconf>=2.3",
    "pillow>=10.0",
    "opencv-python>=4.8",
    "jinja2>=3.1",
    "numpy>=1.24",
    # Phase 2: Multi-modal visualization support
    "open3d>=0.18",
    "plotly>=5.0",
    "numpydantic>=1.6",
    # Phase 3: Performance monitoring and debugging
    "loguru>=0.7",
    "pyarrow>=13.0",
    "psutil>=5.9",
    "scikit-learn>=1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
]

[project.scripts]
viz-art-logs = "viz_art.cli.logs:query_logs_cli"
viz-art-metrics = "viz_art.cli.metrics:view_metrics_cli"
viz-art-validate = "viz_art.cli.validate:validate_cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/viz_art --cov-report=term-missing"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.ruff]
line-length = 88
target-version = "py38"
